+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -3,6 +3,9 @@
 Release History
 ===============
 
+* [BREAKING CHANGE]: previously deprecated command group `network interface-endpoints` has been removed. Use `network private-endpoints` instead.
+* `network vnet subnet create/update`: Added `--nat-gateway` argument for attaching to a NAT gateway.
+
 2.4.0
 +++++
 * [BREAKING CHANGE] `vnet create/update`: Replaced `--cache` arugment with `--defer`.